back_end/requirements.txt
openai
python-dotenv
umap-learn
fastapi
uvicorn